Add new mutation for offer orders
