+++ app/graphql/mutations/create_offer_order_with_artwork.rb
@@ -1,4 +1,4 @@
-class Mutations::CreateOrderWithArtwork < Mutations::BaseMutation
+class Mutations::CreateOfferOrderWithArtwork < Mutations::BaseMutation
   null true
 
   argument :artwork_id, String, 'Artwork Id', required: true
@@ -8,7 +8,7 @@ class Mutations::CreateOrderWithArtwork < Mutations::BaseMutation
   field :order_or_error, Mutations::OrderOrFailureUnionType, 'A union of success/failure', null: false
 
   def resolve(artwork_id:, edition_set_id: nil, quantity: 1)
-    service = CreateBuyOrderService.new(user_id: context[:current_user][:id], artwork_id: artwork_id, edition_set_id: edition_set_id, quantity: quantity)
+    service = CreateOfferOrderService.new(user_id: context[:current_user][:id], artwork_id: artwork_id, edition_set_id: edition_set_id, quantity: quantity)
     service.process!
     {
       order_or_error: { order: service.order }
