+++ app/graphql/types/mutation_type.rb
@@ -1,5 +1,6 @@
 class Types::MutationType < Types::BaseObject
   field :create_order_with_artwork, mutation: Mutations::CreateOrderWithArtwork
+  field :create_offer_order_with_artwork, mutation: Mutations::CreateOfferOrderWithArtwork
   field :set_shipping, mutation: Mutations::SetShipping
   field :set_payment, mutation: Mutations::SetPayment
   field :submit_order, mutation: Mutations::SubmitOrder
