+++ app/services/create_order_service.rb
@@ -40,14 +40,20 @@ class CreateOrderService
     raise Errors::ValidationError.new(:invalid_order, message: e.message)
   end
 
+  protected
+
+  def assert_create!(_artwork)
+    raise NotImplementedError
+  end
+
   private
 
   def pre_process!
     @artwork = GravityService.get_artwork(@artwork_id)
     raise Errors::ValidationError.new(:unknown_artwork, artwork_id: @artwork_id) if @artwork.nil?
     raise Errors::ValidationError.new(:unpublished_artwork, artwork_id: @artwork_id) unless @artwork[:published]
-    raise Errors::ValidationError.new(:not_acquireable, artwork_id: @artwork_id) unless @artwork[:acquireable]
 
+    assert_create!(@artwork)
     find_verify_edition_set
   end
