+++ spec/controllers/api/requests/create_offer_order_with_artwork_mutation_request_spec.rb
@@ -17,8 +17,8 @@ describe Api::GraphqlController, type: :request do
     end
     let(:mutation) do
       <<-GRAPHQL
-        mutation($input: CreateOrderWithArtworkInput!) {
-          createOrderWithArtwork(input: $input) {
+        mutation($input: CreateOfferOrderWithArtworkInput!) {
+          createOfferOrderWithArtwork(input: $input) {
             orderOrError {
               ... on OrderWithMutationSuccess {
                 order {
@@ -67,11 +67,11 @@ describe Api::GraphqlController, type: :request do
         it 'does not create order and returns proper error' do
           expect do
             response = client.execute(mutation, input: mutation_input)
-            expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:order)
-            expect(response.data.create_order_with_artwork.order_or_error.error).not_to be_nil
+            expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:order)
+            expect(response.data.create_offer_order_with_artwork.order_or_error.error).not_to be_nil
 
-            expect(response.data.create_order_with_artwork.order_or_error.error.type).to eq 'validation'
-            expect(response.data.create_order_with_artwork.order_or_error.error.code).to eq 'unknown_artwork'
+            expect(response.data.create_offer_order_with_artwork.order_or_error.error.type).to eq 'validation'
+            expect(response.data.create_offer_order_with_artwork.order_or_error.error.code).to eq 'unknown_artwork'
           end.to change(Order, :count).by(0).and change(LineItem, :count).by(0)
         end
       end
@@ -86,10 +86,9 @@ describe Api::GraphqlController, type: :request do
             it 'uses artworks edition set' do
               expect do
                 response = client.execute(mutation, input: mutation_input.except(:editionSetId))
-                expect(response.data.create_order_with_artwork.order_or_error.order.id).not_to be_nil
-                expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:error)
-                order = Order.find(response.data.create_order_with_artwork.order_or_error.order.id)
-                expect(order.mode).to eq Order::BUY
+                expect(response.data.create_offer_order_with_artwork.order_or_error.order.id).not_to be_nil
+                expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:error)
+                order = Order.find(response.data.create_offer_order_with_artwork.order_or_error.order.id)
                 expect(order.currency_code).to eq 'USD'
                 expect(order.buyer_id).to eq jwt_user_id
                 expect(order.seller_id).to eq partner_id
@@ -106,11 +105,10 @@ describe Api::GraphqlController, type: :request do
             it 'creates order with edition_set price' do
               expect do
                 response = client.execute(mutation, input: mutation_input)
-                expect(response.data.create_order_with_artwork.order_or_error.order.id).not_to be_nil
-                expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:error)
+                expect(response.data.create_offer_order_with_artwork.order_or_error.order.id).not_to be_nil
+                expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:error)
 
-                order = Order.find(response.data.create_order_with_artwork.order_or_error.order.id)
-                expect(order.mode).to eq Order::BUY
+                order = Order.find(response.data.create_offer_order_with_artwork.order_or_error.order.id)
                 expect(order.currency_code).to eq 'USD'
                 expect(order.buyer_id).to eq jwt_user_id
                 expect(order.seller_id).to eq partner_id
@@ -165,11 +163,11 @@ describe Api::GraphqlController, type: :request do
             it 'creates order with edition_set price' do
               expect do
                 response = client.execute(mutation, input: mutation_input)
-                expect(response.data.create_order_with_artwork.order_or_error.order.id).not_to be_nil
-                expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:error)
+                expect(response.data.create_offer_order_with_artwork.order_or_error.order.id).not_to be_nil
+                expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:error)
 
-                order = Order.find(response.data.create_order_with_artwork.order_or_error.order.id)
-                expect(order.mode).to eq Order::BUY
+                order = Order.find(response.data.create_offer_order_with_artwork.order_or_error.order.id)
+                expect(order.mode).to eq Order::OFFER
                 expect(order.currency_code).to eq 'USD'
                 expect(order.buyer_id).to eq jwt_user_id
                 expect(order.seller_id).to eq partner_id
@@ -188,10 +186,10 @@ describe Api::GraphqlController, type: :request do
           it 'creates order with artwork price' do
             expect do
               response = client.execute(mutation, input: mutation_input.except(:editionSetId))
-              expect(response.data.create_order_with_artwork.order_or_error.order.id).not_to be_nil
-              expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:error)
-              order = Order.find(response.data.create_order_with_artwork.order_or_error.order.id)
-              expect(order.mode).to eq Order::BUY
+              expect(response.data.create_offer_order_with_artwork.order_or_error.order.id).not_to be_nil
+              expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:error)
+              order = Order.find(response.data.create_offer_order_with_artwork.order_or_error.order.id)
+              expect(order.mode).to eq Order::OFFER
               expect(order.currency_code).to eq 'USD'
               expect(order.buyer_id).to eq jwt_user_id
               expect(order.seller_id).to eq partner_id
@@ -209,10 +207,10 @@ describe Api::GraphqlController, type: :request do
           it 'creates order with artwork price' do
             expect do
               response = client.execute(mutation, input: mutation_input.except(:editionSetId))
-              expect(response.data.create_order_with_artwork.order_or_error.order.id).not_to be_nil
-              expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:error)
-              order = Order.find(response.data.create_order_with_artwork.order_or_error.order.id)
-              expect(order.mode).to eq Order::BUY
+              expect(response.data.create_offer_order_with_artwork.order_or_error.order.id).not_to be_nil
+              expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:error)
+              order = Order.find(response.data.create_offer_order_with_artwork.order_or_error.order.id)
+              expect(order.mode).to eq Order::OFFER
               expect(order.currency_code).to eq 'USD'
               expect(order.buyer_id).to eq jwt_user_id
               expect(order.seller_id).to eq partner_id
@@ -227,11 +225,11 @@ describe Api::GraphqlController, type: :request do
             it 'defaults to 1' do
               expect do
                 response = client.execute(mutation, input: { artworkId: artwork_id })
-                expect(response.data.create_order_with_artwork.order_or_error.order.id).not_to be_nil
-                expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:error)
+                expect(response.data.create_offer_order_with_artwork.order_or_error.order.id).not_to be_nil
+                expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:error)
 
-                order = Order.find(response.data.create_order_with_artwork.order_or_error.order.id)
-                expect(order.mode).to eq Order::BUY
+                order = Order.find(response.data.create_offer_order_with_artwork.order_or_error.order.id)
+                expect(order.mode).to eq Order::OFFER
                 expect(order.currency_code).to eq 'USD'
                 expect(order.buyer_id).to eq jwt_user_id
                 expect(order.seller_id).to eq partner_id
@@ -252,8 +250,8 @@ describe Api::GraphqlController, type: :request do
             it 'creates a new order' do
               expect do
                 response = client.execute(mutation, input: { artworkId: artwork_id })
-                expect(response.data.create_order_with_artwork.order_or_error.order.id).not_to be_nil
-                expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:error)
+                expect(response.data.create_offer_order_with_artwork.order_or_error.order.id).not_to be_nil
+                expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:error)
                 expect(order.reload.state).to eq Order::PENDING
               end.to change(Order, :count).by(1)
             end
@@ -268,11 +266,11 @@ describe Api::GraphqlController, type: :request do
         it 'returns error' do
           expect do
             response = client.execute(mutation, input: mutation_input.except(:editionSetId))
-            expect(response.data.create_order_with_artwork.order_or_error).not_to respond_to(:order)
-            expect(response.data.create_order_with_artwork.order_or_error.error).not_to be_nil
+            expect(response.data.create_offer_order_with_artwork.order_or_error).not_to respond_to(:order)
+            expect(response.data.create_offer_order_with_artwork.order_or_error.error).not_to be_nil
 
-            expect(response.data.create_order_with_artwork.order_or_error.error.type).to eq 'validation'
-            expect(response.data.create_order_with_artwork.order_or_error.error.code).to eq 'invalid_order'
+            expect(response.data.create_offer_order_with_artwork.order_or_error.error.type).to eq 'validation'
+            expect(response.data.create_offer_order_with_artwork.order_or_error.error.code).to eq 'invalid_order'
           end.to change(Order, :count).by(0)
         end
       end
